quick fix for disciplined groups bug
Currently, deleted groups are set to disciplined and remain in the player's list of groups. If this puts them above the group limit they are unable to create any more groups.

As quick fix for this, I made the statement for checking the amount of groups a player is part of exclude those that are disciplined by either admin intervention or this bug.

I'll work on a long term fix, but didn't want to attempt one without having a server set up with citadel to thoroughly test it. Since this problem is affecting people in game at the moment (see TheJDz's bug report), I figure a simple change that won't need comprehensive testing would be worth it in the mean-time.